+++ CHANGELOG.md
@@ -6,6 +6,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## 1.2.0
+
+### Added
+ - Fetch FTS HPC Data using google sheet.
+ - Add visibility support for project. (Public, Login required, IFRC Only)
+ - New Programme Type `Domestic`
+ - Add Bulk Project Import in Admin Panel.
+ - Enable history for Project changes.
+ - Add Sector/SectorTag `Health (private)` and `COVID-19`.
+ - Add API for Project for region.
+ - Add Multiselect filters for Project API enumfields.
+
+### Changed
+ - Change Sector/SectorTag `Health` to `Health (public)`.
+
 ## 1.1.272
 
 ### Added
@@ -1123,7 +1138,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## 0.1.20
 
-[Unreleased]: https://github.com/IFRCGo/go-api/compare/1.1.272...HEAD
+[Unreleased]: https://github.com/IFRCGo/go-api/compare/1.2.0...HEAD
+[1.2.0]: https://github.com/IFRCGo/go-api/compare/1.2.0...1.1.272
 [1.1.272]: https://github.com/IFRCGo/go-api/compare/1.1.271...1.1.272
 [1.1.271]: https://github.com/IFRCGo/go-api/compare/1.1.269...1.1.271
 [1.1.269]: https://github.com/IFRCGo/go-api/compare/1.1.268...1.1.269